browser/module: fix statement binding
This was tricky to figure out. Using the ENVIRONMENT=web flag for the browser
build started causing prepared statements to fail - it would behave as if
no params were passed to db.exec("...", []) as the second arg.

It turns out that the closure compiler was removing the bind() call..

It's not clear to me why, but adding externs that map out the exported
structure fixes the issue.

Some other notes:
- emscripten runs with ADVANCED mode for closure compiler
- closure compiler reuses variables...
- use --closure-args=--debug and --closure-args=--pretty-print for clues